requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-sqlalchemy==2.5.1
web3==6.0.0
sqlalchemy==1.4.39
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
requests==2.28.1
python-dotenv==0.19.0
plotly==5.13.0